% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SRxgboost_calculate_distance.R
\name{SRxgboost_calculate_distance}
\alias{SRxgboost_calculate_distance}
\title{SRxgboost_calculate_distance}
\usage{
SRxgboost_calculate_distance(variable_old, variable_new, bins = 5)
}
\arguments{
\item{lauf}{character}

\item{rank}{integer}

\item{plots}{boolean}

\item{min_rel_Gain}{numeric}

\item{sample}{integer}

\item{pdp_sample}{integer}

\item{pdp_cuts}{integer}

\item{n_core}{character}

\item{silent}{boolean}
}
\value{
numeric value
}
\description{
Calculates the distance of two vectors.
}
