% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SRxgboost_calculate_drift.R
\name{SRxgboost_calculate_drift}
\alias{SRxgboost_calculate_drift}
\title{SRxgboost_calculate_drift}
\usage{
SRxgboost_calculate_drift(
  data_old,
  data_new,
  n_bins = 10,
  n_plots = 10,
  name = ""
)
}
\arguments{
\item{data_old}{data.frame}

\item{data_new}{data.frame}

\item{n_bins}{integer}

\item{n_plots}{integer}

\item{name}{character}
}
\value{
several files in folder
}
\description{
Calculates drift for example between train and test-data or on residuals.
}
