% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SRxgboost_check_uncertainty.R
\name{SRxgboost_check_uncertainty}
\alias{SRxgboost_check_uncertainty}
\title{SRxgboost_check_uncertainty}
\usage{
SRxgboost_check_uncertainty(lauf)
}
\arguments{
\item{lauf}{character}
}
\value{
uncertainty stats, list with ids and plot
}
\description{
Plots model results for a selected model.
}
