% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SRxgboost_compare_sample_methods.R
\name{SRxgboost_compare_sample_methods}
\alias{SRxgboost_compare_sample_methods}
\title{SRxgboost_compare_sample_methods}
\usage{
SRxgboost_compare_sample_methods(
  df_train,
  df_test = NULL,
  folds,
  runs = 2,
  sample_methods
)
}
\arguments{
\item{df_train}{data.frame}

\item{df_test}{data.frame}

\item{folds}{list, output from SRxgboost_sample_training_data}

\item{runs}{integer}

\item{sample_methods}{list}
}
\value{
lots of files and a graphic of method comparison
}
\description{
Function to compare sample methods with XGBOOST-models.
}
\details{
df_new <- SR_sample_training_data(df = df, folds = folds, sample_method = "ubOver")
df_new <- SR_sample_training_data(df = df, folds = folds, sample_method = "ubOver")
df_new <- SR_sample_training_data(df = df, folds = folds, sample_method = "ubUnder")
df_new <- SR_sample_training_data(df = df, folds = folds, sample_method = "ubSMOTE")
df_new <- SR_sample_training_data(df = df, folds = folds, sample_method = "ubOSS")
df_new <- SR_sample_training_data(df = df, folds = folds, sample_method = "ubCNN")
df_new <- SR_sample_training_data(df = df, folds = folds, sample_method = "ubENN")   # error?
df_new <- SR_sample_training_data(df = df, folds = folds, sample_method = "ubNCL")
df_new <- SR_sample_training_data(df = df, folds = folds, sample_method = "ubTomek")
xx_ubUnder_perc50 <- SR_sample_training_data(df = df, folds = folds,
                                            sample_method = "ubUnder", perc = 50)
xx_ubUnder_perc40 <- SR_sample_training_data(df = df, folds = folds,
                                             sample_method = "ubUnder", perc = 40)
xx_ubSMOTE_po200_pu200_k5 <- SR_sample_training_data(df = df, folds = folds,
                                                     sample_method = "ubSMOTE",
                                                     percOver = 200, percUnder = 200, k = 5)

https://cran.r-project.org/web/packages/unbalanced/unbalanced.pdf
}
