% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SRxgboost_covariate_drift_XGB.R
\name{SRxgboost_covariate_drift_XGB}
\alias{SRxgboost_covariate_drift_XGB}
\title{SRxgboost_covariate_drift_XGB}
\usage{
SRxgboost_covariate_drift_XGB(df1 = NULL, df2 = NULL, pdp_parallel = TRUE)
}
\arguments{
\item{df1}{data.frame}

\item{df2}{data.frame}

\item{pdp_parallel}{boolean, default = TRUE}
}
\value{
several files in folder
}
\description{
Check for covariate drift with XGBOOST Classification-Model (y = 0/1).
}
