% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SRxgboost_create_folds.R
\name{SRxgboost_create_folds}
\alias{SRxgboost_create_folds}
\title{SRxgboost_create_folds}
\usage{
SRxgboost_create_folds(df, foldcolumn, k = 5)
}
\arguments{
\item{df}{data.frame}

\item{foldcolumn}{character}

\item{k}{integer}
}
\value{
list
}
\description{
Creates custom folds for 'SRxgboost_run(folds = ...)', for example time based,
groups, ...
}
\examples{

}
