% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SRxgboost_data_prep.R
\name{SRxgboost_data_prep}
\alias{SRxgboost_data_prep}
\title{SRxgboost_data_prep}
\usage{
SRxgboost_data_prep(
  yname,
  data_train,
  data_test = NULL,
  no_folds = NULL,
  folds = NULL,
  eval_index = NULL,
  nthreads = NULL,
  objective = NULL,
  weights = NULL,
  label_encoding = TRUE,
  add_random_variables = FALSE,
  check_covariate_drift = FALSE
)
}
\arguments{
\item{yname}{character}

\item{data_train}{data.frame}

\item{data_test}{data.frame}

\item{no_folds}{integer}

\item{folds}{list, output from xxx}

\item{eval_index}{vector}

\item{objective}{character: one of "regression", "classification", "multilabel"}

\item{weights}{table of weights (multilabel classification with "weighted_precision" only)}

\item{label_encoding}{boolean}

\item{add_random_variables}{boolean}

\item{check_covariate_drift}{boolean}
}
\value{
assigns output into global workspace
}
\description{
Prepare data for SRxgboost modeling.
Creates data for xgboost modeling function SRxgboost_run (y, partitions,
train, test).
Creates a sparse matrix and dmatrix.
}
