% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SRxgboost_f1.R
\name{SRxgboost_f1}
\alias{SRxgboost_f1}
\title{F1 score}
\usage{
SRxgboost_f1(pred, labels)
}
\arguments{
\item{pred}{numeric vector}

\item{labels}{integer vector 0/1}
}
\value{
list with f1 value und optimal cutoff
}
\description{
The F1 score is the harmonic mean of the precision and recall in binary
classification.
}
\examples{
SRxgboost_f1(pred = seq(0.1, 1, 0.1),
             labels = c(1, 0, 0, 0, 1, 1, 0, 1, 1, 1))

}
