% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SRxgboost_get_OOFforecast_TESTforecast.R
\name{SRxgboost_get_OOFforecast_TESTforecast}
\alias{SRxgboost_get_OOFforecast_TESTforecast}
\title{SRxgboost_get_OOFforecast_TESTforecast}
\usage{
SRxgboost_get_OOFforecast_TESTforecast(lauf, top_rank = 1, ensemble = FALSE)
}
\arguments{
\item{lauf}{character}

\item{top_rank}{integer}

\item{ensemble}{boolean}
}
\value{
predictions as data.frames in memory
}
\description{
Load OOFforecast and TESTforecast from rds-files.
}
