% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SRxgboost_get_summary_CV.R
\name{SRxgboost_get_summary_CV}
\alias{SRxgboost_get_summary_CV}
\title{SRxgboost_get_summary_CV}
\usage{
SRxgboost_get_summary_CV(lauf, ensemble = FALSE)
}
\arguments{
\item{lauf}{character}

\item{ensemble}{boolean}
}
\value{
data.frame
}
\description{
Import the summary of 'SRxgboost_run()'.
}
