% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SRxgboost_h2o_automl.R
\name{SRxgboost_h2o_automl}
\alias{SRxgboost_h2o_automl}
\title{SRxgboost_h2o_automl}
\usage{
SRxgboost_h2o_automl(
  y = NULL,
  data_train = NULL,
  data_test = NULL,
  metric = NULL,
  kfold = 5,
  max_runtime_sec = 5 * 60,
  run_shap = FALSE,
  nthreads = NULL,
  max_mem_size = "16g"
)
}
\arguments{
\item{y}{character}

\item{data_train}{data.frame}

\item{data_test}{data.frame}

\item{metric}{character: "AUTO", "RMSE", "MAE", "RMSLE", "AUC", "AUCPR", "logloss",
"lift_top_group", "misclassification", "mean_per_class_error"}

\item{kfold}{integer}

\item{max_runtime_sec}{integer}

\item{run_shap}{boolean}
}
\value{
writes a lot of files
}
\description{
Runs several models with h2o.automl.
Install latestest version of h2o:
install.packages("h2o", type="source",
                 repos=(c("http://h2o-release.s3.amazonaws.com/h2o/latest_stable_R")))
}
