% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SRxgboost_load_everything.R
\name{SRxgboost_load_everything}
\alias{SRxgboost_load_everything}
\title{SRxgboost_load_everything}
\usage{
SRxgboost_load_everything(lauf, nthreads = NULL)
}
\arguments{
\item{lauf}{character}
}
\value{
loads rds-files into global environment
}
\description{
Loads all previously prepared objects for SRxgboost modeling into global
environment.
}
