% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SRxgboost_make_ensemble.R
\name{SRxgboost_make_ensemble}
\alias{SRxgboost_make_ensemble}
\title{SRxgboost_make_ensemble}
\usage{
SRxgboost_make_ensemble(name, lauf, top_rank = 3)
}
\arguments{
\item{name}{character: folder name for results}

\item{lauf}{character: 1 or several "lauf"-names of previous runs, e.g. c("lauf1", "lauf2")}

\item{top_rank}{integer: 1 integer for each lauf, e.g. c(3, 5)}
}
\value{
forecast and model metrics for ensemble (OOF and TEST if available)
        as rds-file and as data.frame in memory.
}
\description{
Tries out several ensemble models from previous runs and generates predictions
for the best ensemble.
}
