% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mcc.R
\name{SRxgboost_mcc}
\alias{SRxgboost_mcc}
\title{Matthews Correlation Coefficient}
\usage{
SRxgboost_mcc(pred, labels)
}
\arguments{
\item{pred}{numeric vector}

\item{labels}{integer vector 0/1}
}
\value{
list with mcc value und optimal cutoff
}
\description{
"The Best Classification Metric You’ve Never Heard Of"
https://towardsdatascience.com/the-best-classification-metric-youve-never-heard-of-the-matthews-correlation-coefficient-3bf50a2f3e9a
}
\examples{
SRxgboost_mcc(pred = seq(0.1, 1, 0.1),
              labels = c(1, 0, 0, 0, 1, 1, 0, 1, 1, 1))

}
