% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SRxgboost_plots.R
\name{SRxgboost_plots}
\alias{SRxgboost_plots}
\title{SRxgboost_plots}
\usage{
SRxgboost_plots(
  lauf,
  rank = 1,
  plots = TRUE,
  min_rel_Gain = 0.01,
  sample = 1e+05,
  pdp_sample = 20000,
  pdp_cuts = 50,
  n_core = "max",
  silent = FALSE
)
}
\arguments{
\item{lauf}{character}

\item{rank}{integer}

\item{plots}{boolean}

\item{min_rel_Gain}{numeric}

\item{sample}{integer}

\item{pdp_sample}{integer}

\item{pdp_cuts}{integer}

\item{n_core}{character}

\item{silent}{boolean}
}
\value{
several files in folder
}
\description{
Plots model results for a selected model.
}
