% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SRxgboost_run.R
\name{SRxgboost_run}
\alias{SRxgboost_run}
\title{SRxgboost_run}
\usage{
SRxgboost_run(
  nround,
  eta,
  obj,
  metric,
  runs,
  nfold = NULL,
  folds = NULL,
  trees = 1,
  dart = 0,
  tree_method = "auto",
  verbose = 0,
  test_param = FALSE,
  shap = TRUE,
  continue_threshold = 0.1,
  run_final_model = TRUE,
  best_params = NULL,
  max_overfit = -1,
  feat_sel = FALSE,
  Feat_Sel = NULL,
  Feat_Sel_Vars = NULL,
  Selected_Features = FALSE
)
}
\arguments{
\item{folds}{list, output from xxx}

\item{yname}{character}

\item{data_train}{data.frame}

\item{data_test}{data.frame}

\item{no_folds}{integer}

\item{eval_index}{vector}

\item{objective}{character}

\item{label_encoding}{boolean}

\item{check_covariate_drift}{boolean}
}
\value{
assigns output into global workspace
}
\description{
Main function to run XGBOOST models (regression, binary classification,
multiclass classification)
Implemented metrics:
"reg:squarederror"  "rmse", "auc", "rmsle", "mae", "mape"
"reg:logistic"      "error", "auc"
"binary:logistic"   "error", "logloss", "auc", "roc", "qwk_score", "f1_score",
                    "mcc_score"
"multi:softprob"    "merror", "logloss"
"multi:softmax"     "merror", "auc"
"rank:pairwise"     "ndcg"
}
\examples{

}
