% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SRxgboost_sample_training_data.R
\name{SRxgboost_sample_training_data}
\alias{SRxgboost_sample_training_data}
\title{SRxgboost_sample_training_data}
\usage{
SRxgboost_sample_training_data(
  df,
  folds,
  sample_method,
  percOver = 300,
  percUnder = 150,
  k = 5,
  perc = 50,
  method = "percPos",
  w = NULL,
  verbose = TRUE
)
}
\arguments{
\item{df}{data.frame}

\item{folds}{list}

\item{sample_method}{character: ubOver, ubUnder, ubSMOTE, ubOSS, ubCNN, ubENN, ubNCL, ubTomek}

\item{percOver}{integer: parameter used in ubSMOTE}

\item{percUnder}{integer: parameter used in ubSMOTE}

\item{k}{integer: parameter used in ubOver, ubSMOTE, ubCNN, ubENN, ubNCL}

\item{perc}{integer: parameter used in ubUnder}

\item{method}{character; parameter used in ubUnder}

\item{w}{???: parameter used in ubUnder}

\item{verbose}{boolean}
}
\value{
return sampled data and new folds
}
\description{
Function to sample training data with different methods.
https://cran.r-project.org/web/packages/unbalanced/unbalanced.pdf
}
