% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SRxgboost_save_everything.R
\name{SRxgboost_save_everything}
\alias{SRxgboost_save_everything}
\title{SRxgboost_save_everything}
\usage{
SRxgboost_save_everything(lauf)
}
\arguments{
\item{lauf}{character}
}
\value{
saves rds files in lauf-path
}
\description{
Saves all objects for SRxgboost modeling as rds-files.
}
