% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SRxgboost_select_variables.R
\name{SRxgboost_select_variables}
\alias{SRxgboost_select_variables}
\title{SRxgboost_select_variables}
\usage{
SRxgboost_select_variables(lauf_all_variables, threshold_cor = 0.9)
}
\arguments{
\item{lauf_all_variables}{character}

\item{threshold_cor}{numeric (default = 0.9)}
}
\value{
data.frame with relevant variables and their variable importance
}
\description{
Take the results of a previous run and deduct relevant variables:
- by comparing the information gain against the mean of random_numeric,
  random_binary and random_category variables created in 'SRxgboost_data_prep'
- removing from highly correlated variables (default threshold = 0.9 like the
  default value in recipes::step_corr()) the variable with lower information gain
Requirements of previous run with all variables provided at 'lauf_all_variables':
- run the following 3 functions:
  - 'SRxgboost_data_prep(..., add_random_variables = TRUE)'
  - 'SRxgboost_run' and
  - 'SRxgboost_plots'
Returns a data.frame with relevant variables and their variable importance
}
