% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SRxgboost_submission.R
\name{SRxgboost_submission}
\alias{SRxgboost_submission}
\title{SRxgboost_submission}
\usage{
SRxgboost_submission(
  lauf,
  rank = 1,
  rounding = 4,
  id_name,
  y_name,
  y_transformation = NULL
)
}
\arguments{
\item{lauf}{character}

\item{rank}{integer}

\item{rounding}{integer}

\item{id_name}{character}

\item{y_name}{character}

\item{y_transformation}{character}
}
\value{
prediction as csv-file and as data.frame in memory
}
\description{
Get predictions from TESTforecast.csv
}
