% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SRxgboost_calculate_distance.R
\name{SRxgboost_calculate_distance}
\alias{SRxgboost_calculate_distance}
\title{SRxgboost_calculate_distance}
\usage{
SRxgboost_calculate_distance(variable_old, variable_new, bins = 5)
}
\arguments{
\item{variable_old}{character}

\item{variable_new}{character}

\item{bins}{integer}
}
\value{
numeric value
}
\description{
Calculates the distance of two vectors.
}
