% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SRxgboost_compare_sample_methods.R
\name{SRxgboost_compare_sample_methods}
\alias{SRxgboost_compare_sample_methods}
\title{SRxgboost_compare_sample_methods}
\usage{
SRxgboost_compare_sample_methods(
  df_train,
  df_test = NULL,
  folds,
  runs = 2,
  sample_methods,
  pdp_parallel = TRUE
)
}
\arguments{
\item{df_train}{data.frame}

\item{df_test}{data.frame}

\item{folds}{list, output from SRxgboost_create_folds}

\item{runs}{integer, number of xgboost runs}

\item{sample_methods}{list, c("ubOver", "ubUnder", "ubSMOTE", "ubOSS", "ubCNN",
"ubENN", "ubNCL", "ubTomek")}

\item{pdp_parallel}{boolean, default = TRUE}
}
\value{
lots of files and a graphic of method comparison
}
\description{
Function to compare sample methods with XGBOOST-models.
}
