% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SRxgboost_h2o_automl.R
\name{SRxgboost_h2o_automl}
\alias{SRxgboost_h2o_automl}
\title{SRxgboost_h2o_automl}
\usage{
SRxgboost_h2o_automl(
  y = NULL,
  data_train = NULL,
  data_test = NULL,
  metric = NULL,
  kfold = 5,
  max_runtime_sec = 5 * 60,
  run_shap = FALSE
)
}
\arguments{
\item{y}{character}

\item{data_train}{data.frame}

\item{data_test}{data.frame}

\item{metric}{character: "AUC", "RMSE", ...}

\item{kfold}{integer}

\item{max_runtime_sec}{integer}

\item{run_shap}{boolean}
}
\value{
writes a lot of files
}
\description{
Runs several models with h2o.automl.
Install latestest version of h2o:
install.packages("h2o", type="source",
}
