% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SRxgboost_make_ensemble.R
\name{SRxgboost_make_ensemble}
\alias{SRxgboost_make_ensemble}
\title{SRxgboost_make_ensemble}
\usage{
SRxgboost_make_ensemble(lauf, top_rank = 3)
}
\arguments{
\item{lauf}{character}

\item{top_rank}{integer}
}
\value{
forecast and model metrics for ensemble (OOF and TEST if available)
        as rds-file and as data.frame in memory
}
\description{
Make an ensemble model and predictions from previous runs.
}
