% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SRxgboost_plots.R
\name{SRxgboost_plots}
\alias{SRxgboost_plots}
\title{SRxgboost_plots}
\usage{
SRxgboost_plots(
  lauf,
  rank = 1,
  plots = TRUE,
  silent = FALSE,
  uncertainty_quantil = NULL,
  sample = 1e+05,
  pdp_sample = 50000,
  pdp_int_sample = 30000,
  pdp_plots = TRUE,
  pdp_parallel = TRUE,
  pdp_min_rel_Gain = 0.01,
  pdp_cuts = NULL,
  pdp_int_cuts_sample = 50
)
}
\arguments{
\item{lauf}{character}

\item{rank}{integer, default = 1}

\item{plots}{boolean, default = TRUE}

\item{silent}{boolean, default = FALSE}

\item{sample}{integer, default = 100000}

\item{pdp_sample}{integer, default = 50000}

\item{pdp_int_sample}{integer, default = 30000}

\item{pdp_plots}{boolean, default = TRUE}

\item{pdp_parallel}{boolean, default = TRUE}

\item{pdp_min_rel_Gain}{numeric, default =  0.01}

\item{pdp_cuts}{integer, default = NULL, depending on number of data points
(max. 50 but in average at least pdp_int_cuts_sample per cut)}

\item{pdp_int_cuts_sample}{integer, default = 50}
}
\value{
several files in folder
}
\description{
Plots model results for a selected model.
}
