% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SRxgboost_plots.R
\name{SRxgboost_plots}
\alias{SRxgboost_plots}
\title{SRxgboost_plots}
\usage{
SRxgboost_plots(
  lauf,
  rank = 1,
  plots = TRUE,
  sample = 1e+05,
  silent = FALSE,
  pdp_plots = TRUE,
  pdp_min_rel_Gain = 0.01,
  pdp_sample = 20000,
  pdp_cuts = NULL,
  pdp_int_sample = 1000,
  pdp_int_cuts_sample = 20,
  pdp_parallel = FALSE
)
}
\arguments{
\item{lauf}{character}

\item{rank}{integer, default = 1}

\item{plots}{boolean, default = TRUE}

\item{sample}{integer, default =}

\item{silent}{boolean, default = FALSE}

\item{pdp_plots}{boolean, default = TRUE}

\item{pdp_min_rel_Gain}{numeric, default =  0.01}

\item{pdp_sample}{integer, default = 20000}

\item{pdp_cuts}{integer, default = NULL, depending on number of data points}

\item{pdp_int_sample}{integer, default = 1000}

\item{pdp_int_cuts_sample}{integer, default = 20}

\item{pdp_parallel}{boolean, default =  = FALSE}
}
\value{
several files in folder
}
\description{
Plots model results for a selected model.
}
