% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SRxgboost_run.R
\name{SRxgboost_run}
\alias{SRxgboost_run}
\title{SRxgboost_run}
\usage{
SRxgboost_run(
  nround,
  eta,
  obj,
  metric,
  runs,
  nfold = NULL,
  folds = NULL,
  trees = 1,
  dart = 0,
  tree_method = "auto",
  verbose = 0,
  test_param = FALSE,
  shap = TRUE,
  continue_threshold = 0.1,
  run_final_model = TRUE,
  best_params = NULL,
  max_overfit = -1,
  feat_sel = FALSE,
  Feat_Sel = NULL,
  Feat_Sel_Vars = NULL,
  Selected_Features = FALSE
)
}
\arguments{
\item{nround}{integer}

\item{eta}{numeric}

\item{obj}{character}

\item{metric}{character}

\item{runs}{integer}

\item{nfold}{integer}

\item{folds}{list, output of SRxgboost_create_folds()}

\item{trees}{integer}

\item{dart}{numeric}

\item{tree_method}{character}

\item{verbose}{integer 0 to 3}

\item{test_param}{boolean}

\item{shap}{boolean}

\item{continue_threshold}{numeric}

\item{run_final_model}{boolean}

\item{best_params}{character}

\item{max_overfit}{numeric between for example -1 (no restriction) or -0.5}

\item{feat_sel}{boolean}

\item{Feat_Sel}{data.frame}

\item{Feat_Sel_Vars}{list}

\item{Selected_Features}{boolean}
}
\value{
several files in folder
}
\description{
Main function to run XGBOOST models (regression, binary classification,
multiclass classification)
Implemented metrics:
"reg:squarederror"  "rmse", "auc", "rmsle", "mae", "mape"
"reg:logistic"      "error", "auc"
"binary:logistic"   "error", "logloss", "auc", "roc", "qwk_score", "f1_score",
                    "mcc_score"
"multi:softprob"    "merror", "logloss"
"multi:softmax"     "merror", "auc"
"rank:pairwise"     "ndcg"
}
