% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SRxgboost_run.R
\name{SRxgboost_run}
\alias{SRxgboost_run}
\title{SRxgboost_run}
\usage{
SRxgboost_run(
  nround = 1000,
  eta = 0.1,
  obj,
  metric,
  runs = 2,
  nfold = NULL,
  folds = NULL,
  early_stopping_rounds = 30,
  scale_pos_weight = 1,
  trees = 1,
  tree_method = "auto",
  verbose = 0,
  test_param = FALSE,
  nthreads = NULL,
  shap = TRUE,
  continue_threshold = 0.1,
  run_final_model = TRUE,
  best_params = NULL,
  max_overfit = -1,
  feat_sel = FALSE,
  Feat_Sel = NULL,
  Feat_Sel_Vars = NULL,
  Selected_Features = FALSE
)
}
\arguments{
\item{nround}{integer}

\item{eta}{numeric}

\item{obj}{character}

\item{metric}{character}

\item{runs}{integer}

\item{nfold}{integer}

\item{folds}{list, output of SRxgboost_create_folds()}

\item{early_stopping_rounds}{integer}

\item{scale_pos_weight}{numeric, control the balance of positive and negative
cases in for unbalanced classification;
typical value: sum(y == 0) / sum(y == 1)}

\item{trees}{integer}

\item{tree_method}{character}

\item{verbose}{integer 0 to 3}

\item{test_param}{boolean}

\item{nthreads}{integer}

\item{shap}{boolean}

\item{continue_threshold}{numeric}

\item{run_final_model}{boolean}

\item{best_params}{character}

\item{max_overfit}{numeric between for example -1 (no restriction) or -0.5}

\item{feat_sel}{boolean}

\item{Feat_Sel}{data.frame}

\item{Feat_Sel_Vars}{list}

\item{Selected_Features}{boolean}
}
\value{
several files in folder
}
\description{
Main function to run XGBOOST models (regression, binary classification,
multiclass classification).
Implemented metrics:
\itemize{
  \item "reg:squarederror": "rmse", "auc", "rmsle", "mae", "mape"
  \item "reg:logistic": "error", "auc"
  \item "binary:logistic": "error", "logloss", "auc", "roc", "qwk_score",
                           "f1_score", "mcc_score", "tpr", "prAUC"
  \item "multi:softprob": "merror", "mlogloss", "mAUC", "weighted_precision"
  \item "multi:softmax": "merror", "mlogloss", "mAUC", "weighted_precision"
  \item "rank:pairwise": "ndcg"
}
}
