% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SRxgboost_sample_training_data.R
\name{SRxgboost_sample_training_data}
\alias{SRxgboost_sample_training_data}
\title{SRxgboost_sample_training_data}
\usage{
SRxgboost_sample_training_data(
  df,
  folds,
  sample_method,
  percOver = 200,
  percUnder = 200,
  k = 5,
  perc = 50,
  method = "percPos",
  w = NULL,
  verbose = TRUE
)
}
\arguments{
\item{df}{data.frame}

\item{folds}{list, output from xxx}

\item{sample_method}{character: ubOver, ubUnder, ubSMOTE, ubOSS, ubCNN, ubENN, ubNCL, ubTomek}

\item{percOver}{integer: parameter used in ubSMOTE}

\item{percUnder}{integer: parameter used in ubSMOTE}

\item{k}{integer: parameter used in ubOver, ubSMOTE, ubCNN, ubENN, ubNCL}

\item{perc}{integer: parameter used in ubUnder}

\item{method}{character; parameter used in ubUnder}

\item{verbose}{boolean}

\item{w:}{parameter used in ubUnder}
}
\value{
return sampled data and new folds
}
\description{
Function to sample training data with different methods.
}
\details{
df_new <- SR_sample_training_data(df = df, folds = folds, sample_method = "ubOver")
df_new <- SR_sample_training_data(df = df, folds = folds, sample_method = "ubOver")
df_new <- SR_sample_training_data(df = df, folds = folds, sample_method = "ubUnder")
df_new <- SR_sample_training_data(df = df, folds = folds, sample_method = "ubSMOTE")
df_new <- SR_sample_training_data(df = df, folds = folds, sample_method = "ubOSS")
df_new <- SR_sample_training_data(df = df, folds = folds, sample_method = "ubCNN")
df_new <- SR_sample_training_data(df = df, folds = folds, sample_method = "ubENN")   # error?
df_new <- SR_sample_training_data(df = df, folds = folds, sample_method = "ubNCL")
df_new <- SR_sample_training_data(df = df, folds = folds, sample_method = "ubTomek")
xx_ubUnder_perc50 <- SR_sample_training_data(df = df, folds = folds,
                                            sample_method = "ubUnder", perc = 50)
xx_ubUnder_perc40 <- SR_sample_training_data(df = df, folds = folds,
                                             sample_method = "ubUnder", perc = 40)
xx_ubSMOTE_po200_pu200_k5 <- SR_sample_training_data(df = df, folds = folds,
                                                     sample_method = "ubSMOTE",
                                                     percOver = 200, percUnder = 200, k = 5)

https://cran.r-project.org/web/packages/unbalanced/unbalanced.pdf
}
